requirements.txt
# requriments
jsonschema
keras-nlp

# optional
google-generativeai # if you use gemini.py
